create slider componet
