--- src/feature/auth/index.tsx
@@ -1,7 +1,0 @@
-export * from "./components";
-export * from "./models";
-export * from "./api";
-export * from "./constants";
-export * from "./hooks";
-export * from "./utils";
-export * from "./enums";
