--- src/feature/auth/models/index.tsx
@@ -1,1 +1,0 @@
-export * from "./button.type";
