--- src/feature/home/components/index.tsx
@@ -1,2 +1,0 @@
-export * from "./category-gallery/cattegory-gallery.component";
-export * from "./site-intro/site-intro.component";
