--- src/feature/home/enums/index.tsx
@@ -1,14 +1,0 @@
-export const BRAND = {
-  value: 200,
-  text: "International Brands",
-};
-
-export const PRODUCTS = {
-  value: 2000,
-  text: "High-Quality Products",
-};
-
-export const CUSTOMERS = {
-  value: 30000,
-  text: "Happy Customers",
-};
