--- src/feature/home/models/cattegory-gallery.type.ts
@@ -1,9 +1,0 @@
-export interface CategoryGalleryProps {
-  galleryItems: SingleCategoryGalleryItem[];
-}
-export interface SingleCategoryGalleryItem {
-  name?: string;
-  image: string;
-  url?: string;
-  slug?: string;
-}
