--- src/feature/home/models/index.tsx
@@ -1,1 +1,0 @@
-export * from "./cattegory-gallery.type";
