--- src/feature/product-category/models/button.type.ts
@@ -1,3 +1,0 @@
-export interface ButtonProps {
-  Label: string;
-}
