--- src/feature/product-category/models/index.tsx
@@ -1,1 +1,0 @@
-export * from "./button.type";
