--- src/feature/shared/api/index.tsx
@@ -1,2 +1,0 @@
-export * from "./api-client";
-export * from "./search.api";
