--- src/feature/shared/components/index.tsx
@@ -1,33 +1,0 @@
-import Button from "./button-component/button.component";
-import Card from "./card-component/card.component";
-import CustomImage from "./image-component/image.component";
-import Price from "./price-component/price.component";
-import ProductCarousel from "./product-carousel-component/product-carousel.component";
-import ProductCounter from "./number-of-product-component/number-of-product.component";
-import Rating from "./rating-component/rating.component";
-import Search from "./search-component/search.component";
-import SearchResult from "./search-component/search-result.component";
-import ShimmerPageLoader, {
-  ShimmerCardLoader,
-  ShimmerContentLoader,
-  ShimmerImageLoader,
-} from "./shimmer-effect/shimmer-effect.component";
-import Slider from "./slider/slider.component";
-import Title from "./title-component/title.component";
-export {
-  Button,
-  Card,
-  CustomImage,
-  Price,
-  ProductCarousel,
-  ProductCounter,
-  Rating,
-  Search,
-  SearchResult,
-  ShimmerCardLoader,
-  ShimmerContentLoader,
-  ShimmerImageLoader,
-  ShimmerPageLoader,
-  Slider,
-  Title,
-};
