--- src/feature/shared/enums/index.tsx
@@ -1,23 +1,0 @@
-export enum BorderRound {
-  FULL = "rounded-full",
-  "3XL" = "rounded-3xl",
-  "2XL" = "rounded-2xl",
-  XL = "rounded-xl",
-  LG = "rounded-lg",
-  MD = "rounder-md",
-  SM = "rounded-sm",
-}
-
-export enum TextHorizontalAlignment {
-  LEFT = "left-0 text-left",
-  CENTER = "left-1/2 transform -translate-x-1/2 text-center",
-  RIGHT = "right-0 text-right",
-}
-export enum TextVerticalAlignment {
-  TOP = "top-0",
-  MIDDLE = "top-1/2 transform -translate-y-1/2",
-  BOTTOM = "bottom-0",
-}
-export enum QueryTime {
-  STALE = 1000 * 60 * 60 * 4,
-}
