--- src/feature/shared/models/button.type.ts
@@ -1,12 +1,0 @@
-import { BorderRound } from "../enums";
-
-export interface ButtonProps extends React.PropsWithChildren {
-  label: string;
-  rounded?: keyof typeof BorderRound;
-  width: string;
-  backgroundColor?: string;
-  color?: string;
-  customStyles?: string;
-  extraClasses?: string;
-  onClick: () => void;
-}
