--- src/feature/shared/models/card.type.ts
@@ -1,10 +1,0 @@
-export interface CardProps {
-  id: string;
-  title?: string;
-  imageURL?: string;
-  rating?: number;
-  price?: number;
-  discount?: number;
-  customStyles?: string;
-  onClick?: (id: string) => void;
-}
