--- src/feature/shared/models/index.tsx
@@ -1,9 +1,0 @@
-export * from "./button.type";
-export * from "./card.type";
-export * from "./image.type";
-export * from "./price.type";
-export * from "./product-carousel.type";
-export * from "./product.type";
-export * from "./rating.type";
-export * from "./search.type";
-export * from "./title.type";
