--- src/feature/shared/models/price.type.ts
@@ -1,4 +1,0 @@
-export interface PriceProps {
-  price: number;
-  discount?: number;
-}
