--- src/feature/shared/models/product-carousel.type.ts
@@ -1,9 +1,0 @@
-import { CardProps } from "./card.type";
-
-export interface ProductCarouselProps extends React.PropsWithChildren {
-  products: CardProps[];
-  carouselTitle: string;
-  showMoreButton?: boolean;
-  buttonText?: string;
-  onClickButton?: () => void;
-}
