--- src/feature/shared/models/product.type.ts
@@ -1,42 +1,0 @@
-export interface ProductResponse {
-  id: string;
-  title: string;
-  description: string;
-  category: string;
-  price: number;
-  discountPercentage: number;
-  rating: number;
-  stock: number;
-  tags: string[];
-  brand: string;
-  sku: string;
-  weight: number;
-  dimensions: ProductDimension[];
-  warrantyInformation: string;
-  shippingInformation: string;
-  availabilityStatus: string;
-  reviews: ProductReview[];
-  returnPolicy: string;
-  minimumOrderQuantity: number;
-  meta: ProductMeta;
-  thumbnail: string;
-  images: string[];
-}
-export interface ProductDimension {
-  width: number;
-  height: number;
-  depth: number;
-}
-export interface ProductReview {
-  rating: number;
-  comment: string;
-  date: string;
-  reviewerName: string;
-  reviewerEmail: string;
-}
-export interface ProductMeta {
-  createdAt: string;
-  updatedAt: string;
-  barcode: string;
-  qrCode: string;
-}
