--- src/feature/shared/models/rating.type.ts
@@ -1,4 +1,0 @@
-export interface RatingProps {
-  rating: number;
-  maxRating: number;
-}
