--- src/feature/shared/models/title.type.ts
@@ -1,6 +1,0 @@
-export interface TitleProps extends React.PropsWithChildren {
-  font?: string;
-  fontSize?: string;
-  color?: string;
-  extraClasses?: string;
-}
