+++ src/main.tsx
@@ -3,7 +3,6 @@ import { createRoot } from 'react-dom/client'
 import App from './App.tsx'
 import './index.css'
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import { SearchResult } from "./feature/index.tsx";
 import Demo from "./demo/demo.tsx";
 
 const router = createBrowserRouter([
@@ -14,10 +13,6 @@ const router = createBrowserRouter([
       {
         path: "/",
         element: <Demo />,
-      },
-      {
-        path: "/search",
-        element: <SearchResult />,
       },
     ],
   },
