+++ src/slider/api-client.ts
