+++ src/slider/slider.css
