+++ src/slider/slider.type.ts
